# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BidOps API
  version: 0.1.0
  summary: BidOps MVP OpenAPI 1차 초안
  description: >
    공공/민간 입찰·제안 업무를 위한 근거 기반 RFP 분석 및 수주 운영 시스템 API.
    문서 업로드 → 비동기 분석 → 요구사항 검토 → 체크리스트/질의/산출물 관리 흐름을 기준으로 정의한다.
servers:
  - url: /api/v1
    description: API v1
...
security:
  - bearerAuth: []
paths:
...
  /projects/{projectId}/analysis-jobs:
    get:
      tags: [AnalysisJobs]
      summary: 프로젝트 기준 최신 Job 목록
      operationId: listAnalysisJobs
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - in: query
          name: job_type
          schema: { $ref: '#/components/schemas/AnalysisJobType' }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/AnalysisJobStatus' }
      responses:
        '200':
          description: 분석 Job 목록
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnalysisJobListResponse' }
    post:
      tags: [AnalysisJobs]
      summary: 분석 시작
      operationId: createAnalysisJob
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnalysisJobCreateRequest' }
      responses:
        '201':
          description: 분석 Job 생성
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnalysisJobResponse' }
  /projects/{projectId}/analysis-jobs/{jobId}:
    get:
      tags: [AnalysisJobs]
      summary: Job 상태 조회
      operationId: getAnalysisJob
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: 분석 Job 상세
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnalysisJobResponse' }
...
  /projects/{projectId}/checklists:
    get:
      tags: [Checklists]
      summary: 체크리스트 목록 조회
      operationId: listChecklistItems
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - in: query
          name: type
          schema: { $ref: '#/components/schemas/ChecklistType' }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/ChecklistStatus' }
        - in: query
          name: risk_level
          schema: { $ref: '#/components/schemas/RiskLevel' }
        - in: query
          name: owner_user_id
          schema: { type: string }
        - in: query
          name: mandatory
          schema: { type: boolean }
        - in: query
          name: keyword
          schema: { type: string, description: "항목 텍스트 검색 (LIKE)" }
        - in: query
          name: requirement_id
          schema: { type: string }
      responses:
        '200':
          description: 체크리스트 목록
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChecklistItemListResponse' }
    post:
      tags: [Checklists]
      summary: 체크리스트 항목 생성
      operationId: createChecklistItem
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChecklistItemCreateRequest' }
      responses:
        '201':
          description: 체크리스트 항목 생성 완료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChecklistItemResponse' }
  /projects/{projectId}/checklists/{checklistItemId}:
    patch:
      tags: [Checklists]
      summary: 체크리스트 항목 수정
      operationId: updateChecklistItem
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ChecklistItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChecklistItemUpdateRequest' }
      responses:
        '200':
          description: 체크리스트 항목 수정 완료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChecklistItemResponse' }
...
  /projects/{projectId}/artifacts:
    get:
      tags: [Artifacts]
      summary: 산출물 목록 조회
      operationId: listArtifacts
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 산출물 목록
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtifactListResponse' }
    post:
      tags: [Artifacts]
      summary: 산출물 생성
      operationId: createArtifact
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ArtifactCreateRequest' }
      responses:
        '201':
          description: 산출물 생성 완료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtifactResponse' }
  /projects/{projectId}/artifacts/{artifactId}:
    get:
      tags: [Artifacts]
      summary: 산출물 상세
      operationId: getArtifact
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ArtifactId'
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtifactResponse' }
    patch:
      tags: [Artifacts]
      summary: 산출물 수정
      operationId: updateArtifact
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ArtifactId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ArtifactUpdateRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtifactResponse' }
    delete:
      tags: [Artifacts]
      summary: 산출물 삭제
      operationId: deleteArtifact
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ArtifactId'
      responses:
        '200':
          description: 삭제 완료
  /projects/{projectId}/artifacts/{artifactId}/status:
    post:
      tags: [Artifacts]
      summary: 산출물 상태 변경
      operationId: changeArtifactStatus
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ArtifactId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/ArtifactStatus' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtifactResponse' }
  /projects/{projectId}/artifacts/{artifactId}/versions:
    get:
      tags: [Artifacts]
      summary: 버전 목록
      operationId: listArtifactVersions
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ArtifactId'
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtifactVersionListResponse' }
    post:
      tags: [Artifacts]
      summary: 버전 업로드
      operationId: uploadArtifactVersion
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ArtifactId'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                version_note: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtifactVersionResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      in: path
      name: projectId
      required: true
      schema: { type: string }
...
    JobId:
      in: path
      name: jobId
      required: true
      schema: { type: string }
...
    ChecklistItemId:
      in: path
      name: checklistItemId
      required: true
      schema: { type: string }
...
    ArtifactId:
      in: path
      name: artifactId
      required: true
      schema: { type: string }
...
  schemas:
    Meta:
      type: object
      properties:
        page: { type: integer }
        size: { type: integer }
        total_count: { type: integer }
      additionalProperties: true
...
    AnalysisJobStatus:
      type: string
      enum: [PENDING, RUNNING, COMPLETED, FAILED]
...
    ChecklistStatus:
      type: string
      enum: [TODO, IN_PROGRESS, DONE, BLOCKED]
    RiskLevel:
      type: string
      enum: [NONE, LOW, MEDIUM, HIGH]
...
    AnalysisJobType:
      type: string
      enum: [OCR, PARSE, REQUIREMENT_EXTRACTION, CHECKLIST_BUILD, QUERY_BUILD, ARTIFACT_BUILD, SEARCH_INDEX, RFP_PARSE]
    ChecklistType:
      type: string
      enum: [SUBMISSION, EVALUATION, EVIDENCE]
    AssetType:
      type: string
      enum: [PROPOSAL, DESIGN, PLAN, REPORT, EVIDENCE, PRESENTATION, ETC]
      description: 산출물 유형
    ArtifactStatus:
      type: string
      enum: [DRAFT, IN_PROGRESS, REVIEW, APPROVED, SUBMITTED]
      description: 산출물 상태
...
    AnalysisJob:
      type: object
      required: [id, project_id, document_id, job_type, status]
      properties:
        id: { type: string }
        project_id: { type: string }
        document_id: { type: string }
        job_type: { $ref: '#/components/schemas/AnalysisJobType' }
        status: { $ref: '#/components/schemas/AnalysisJobStatus' }
        progress: { type: integer, minimum: 0, maximum: 100 }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        error_code: { type: string }
        error_message: { type: string }
...
    ChecklistItem:
      type: object
      required: [id, item_code, item_text, mandatory_flag, current_status, risk_level]
      properties:
        id: { type: string }
        checklist_type: { $ref: '#/components/schemas/ChecklistType' }
        linked_requirement_id: { type: string }
        source_excerpt_id: { type: string }
        owner_user_id: { type: string }
        item_code: { type: string }
        item_text: { type: string }
        description: { type: string }
        mandatory_flag: { type: boolean }
        current_status: { $ref: '#/components/schemas/ChecklistStatus' }
        risk_level: { $ref: '#/components/schemas/RiskLevel' }
        risk_note: { type: string }
        due_date: { type: string, format: date }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Artifact:
      type: object
      required: [id, project_id, title, asset_type, status]
      properties:
        id: { type: string }
        project_id: { type: string }
        title: { type: string }
        asset_type: { $ref: '#/components/schemas/AssetType' }
        status: { $ref: '#/components/schemas/ArtifactStatus' }
        description: { type: string }
        linked_requirement_id: { type: string }
        linked_checklist_item_id: { type: string }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
        updated_at: { type: string, format: date-time }
    ArtifactVersion:
      type: object
      required: [id, version, file_name]
      properties:
        id: { type: string }
        version: { type: integer }
        file_name: { type: string }
        version_note: { type: string }
        uploaded_by: { type: string }
        viewer_url: { type: string, description: 파일 열기/다운로드 URL }
        created_at: { type: string, format: date-time }
...
    AnalysisJobCreateRequest:
      type: object
      required: [document_id, job_type]
      properties:
        document_id: { type: string }
        job_type: { $ref: '#/components/schemas/AnalysisJobType' }
...
    ChecklistItemCreateRequest:
      type: object
      required: [checklist_type, item_text, mandatory_flag]
      properties:
        checklist_type: { $ref: '#/components/schemas/ChecklistType' }
        linked_requirement_id: { type: string }
        source_excerpt_id: { type: string }
        owner_user_id: { type: string }
        item_code: { type: string }
        item_text: { type: string }
        description: { type: string }
        mandatory_flag: { type: boolean }
        due_date: { type: string, format: date }
    ChecklistItemUpdateRequest:
      type: object
      properties:
        owner_user_id: { type: string }
        description: { type: string }
        current_status: { $ref: '#/components/schemas/ChecklistStatus' }
        risk_level: { $ref: '#/components/schemas/RiskLevel' }
        risk_note: { type: string }
        due_date: { type: string, format: date }
...
    ArtifactCreateRequest:
      type: object
      required: [title, asset_type]
      properties:
        title: { type: string }
        asset_type: { $ref: '#/components/schemas/AssetType' }
        description: { type: string }
        linked_requirement_id: { type: string }
        linked_checklist_item_id: { type: string }
    ArtifactUpdateRequest:
      type: object
      properties:
        title: { type: string }
        asset_type: { $ref: '#/components/schemas/AssetType' }
        description: { type: string }
        linked_requirement_id: { type: string }
        linked_checklist_item_id: { type: string }
...
    AnalysisJobResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/AnalysisJob' }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { type: 'null' }
    AnalysisJobListResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, const: true }
        data:
          type: object
          required: [items, total_count]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/AnalysisJob' }
            total_count: { type: integer }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { type: 'null' }
...
    ChecklistItemResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/ChecklistItem' }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { type: 'null' }
    ChecklistItemListResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, const: true }
        data:
          type: object
          required: [items, total_count]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/ChecklistItem' }
            total_count: { type: integer }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { type: 'null' }
...
    ArtifactResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/Artifact' }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { type: 'null' }
    ArtifactListResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, const: true }
        data:
          type: object
          required: [items, total_count]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Artifact' }
            total_count: { type: integer }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { type: 'null' }
    ArtifactVersionResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/ArtifactVersion' }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { type: 'null' }
    ArtifactVersionListResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, const: true }
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/ArtifactVersion' }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { type: 'null' }
...